Remove service worker logger
