+++ components/serviceworker.js
@@ -1,7 +1,6 @@
 import { createContext, useContext, useEffect, useState, useCallback, useMemo } from 'react'
 import { Workbox } from 'workbox-window'
 import { gql, useMutation } from '@apollo/client'
-import { detectOS, useServiceWorkerLogger } from './logger'
 
 const applicationServerKey = process.env.NEXT_PUBLIC_VAPID_PUBKEY
 
@@ -44,7 +43,6 @@ export const ServiceWorkerProvider = ({ children }) => {
           }
         }
       `)
-  const logger = useServiceWorkerLogger()
 
   // I am not entirely sure if this is needed since at least in Brave,
   // using `registration.pushManager.subscribe` also prompts the user.
@@ -77,7 +75,6 @@ export const ServiceWorkerProvider = ({ children }) => {
     // see https://stackoverflow.com/a/69624651
     let pushSubscription = await registration.pushManager.subscribe(subscribeOptions)
     const { endpoint } = pushSubscription
-    logger.info('subscribed to push notifications', { endpoint })
     // convert keys from ArrayBuffer to string
     pushSubscription = JSON.parse(JSON.stringify(pushSubscription))
     // Send subscription to service worker to save it so we can use it later during `pushsubscriptionchange`
@@ -86,7 +83,6 @@ export const ServiceWorkerProvider = ({ children }) => {
       action: STORE_SUBSCRIPTION,
       subscription: pushSubscription
     })
-    logger.info('sent STORE_SUBSCRIPTION to service worker', { endpoint })
     // send subscription to server
     const variables = {
       endpoint,
@@ -94,18 +90,15 @@ export const ServiceWorkerProvider = ({ children }) => {
       auth: pushSubscription.keys.auth
     }
     await savePushSubscription({ variables })
-    logger.info('sent push subscription to server', { endpoint })
   }
 
   const unsubscribeFromPushNotifications = async (subscription) => {
     await subscription.unsubscribe()
     const { endpoint } = subscription
-    logger.info('unsubscribed from push notifications', { endpoint })
     await deletePushSubscription({ variables: { endpoint } })
     // also delete push subscription in IndexedDB so we can tell if the user disabled push subscriptions
     // or we lost the push subscription due to a bug
     navigator.serviceWorker.controller.postMessage({ action: DELETE_SUBSCRIPTION })
-    logger.info('deleted push subscription from server', { endpoint })
   }
 
   const togglePushSubscription = useCallback(async () => {
@@ -117,9 +110,7 @@ export const ServiceWorkerProvider = ({ children }) => {
       // request persistent storage: https://web.dev/learn/pwa/offline-data#data_persistence
       const persisted = await navigator?.storage?.persisted?.()
       if (!persisted && navigator?.storage?.persist) {
-        return navigator.storage.persist().then(persistent => {
-          logger.info('persistent storage:', persistent)
-        }).catch(logger.error)
+        return navigator.storage.persist()
       }
     })
   })
@@ -133,13 +124,11 @@ export const ServiceWorkerProvider = ({ children }) => {
     setPermission({ notification: 'Notification' in window ? window.Notification.permission : 'denied' })
 
     if (!('serviceWorker' in navigator)) {
-      logger.info('device does not support service worker')
       return
     }
 
     const wb = new Workbox('/sw.js', { scope: '/' })
     wb.register().then(registration => {
-      logger.info('service worker registration successful')
       setRegistration(registration)
     })
   }, [])
@@ -159,9 +148,7 @@ export const ServiceWorkerProvider = ({ children }) => {
     // we sync with server manually by checking on every page reload if the push subscription changed.
     // see https://medium.com/@madridserginho/how-to-handle-webpush-api-pushsubscriptionchange-event-in-modern-browsers-6e47840d756f
     navigator?.serviceWorker?.controller?.postMessage?.({ action: STORE_OS, os: detectOS() })
-    logger.info('sent STORE_OS to service worker: ', detectOS())
     navigator?.serviceWorker?.controller?.postMessage?.({ action: SYNC_SUBSCRIPTION })
-    logger.info('sent SYNC_SUBSCRIPTION to service worker')
   }, [registration, permission.notification])
 
   const contextValue = useMemo(() => ({
@@ -177,6 +164,31 @@ export const ServiceWorkerProvider = ({ children }) => {
       {children}
     </ServiceWorkerContext.Provider>
   )
+}
+
+function detectOS () {
+  if (!window.navigator) return ''
+
+  const userAgent = window.navigator.userAgent
+  const platform = window.navigator.userAgentData?.platform || window.navigator.platform
+  const macosPlatforms = ['Macintosh', 'MacIntel', 'MacPPC', 'Mac68K']
+  const windowsPlatforms = ['Win32', 'Win64', 'Windows', 'WinCE']
+  const iosPlatforms = ['iPhone', 'iPad', 'iPod']
+  let os = null
+
+  if (macosPlatforms.indexOf(platform) !== -1) {
+    os = 'Mac OS'
+  } else if (iosPlatforms.indexOf(platform) !== -1) {
+    os = 'iOS'
+  } else if (windowsPlatforms.indexOf(platform) !== -1) {
+    os = 'Windows'
+  } else if (/Android/.test(userAgent)) {
+    os = 'Android'
+  } else if (/Linux/.test(platform)) {
+    os = 'Linux'
+  }
+
+  return os
 }
 
 export function useServiceWorker () {
