+++ lib/validate.js
@@ -472,7 +472,6 @@ export const settingsSchema = object().shape({
   hideNostr: boolean(),
   hideTwitter: boolean(),
   hideWalletBalance: boolean(),
-  diagnostics: boolean(),
   noReferralLinks: boolean(),
   hideIsContributor: boolean(),
   disableFreebies: boolean().nullable(),
