+++ pages/_app.js
@@ -16,7 +16,6 @@ import { ServiceWorkerProvider } from '@/components/serviceworker'
 import { SSR } from '@/lib/constants'
 import NProgress from 'nprogress'
 import 'nprogress/nprogress.css'
-import { LoggerProvider } from '@/components/logger'
 import { ChainFeeProvider } from '@/components/chain-fee.js'
 import dynamic from 'next/dynamic'
 import { HasNewNotesProvider } from '@/components/use-has-new-notes'
@@ -114,28 +113,26 @@ export default function MyApp ({ Component, pageProps: { ...props } }) {
             <MeProvider me={me}>
               <WalletsProvider>
                 <HasNewNotesProvider>
-                  <LoggerProvider>
-                    <WebLnProvider>
-                      <ServiceWorkerProvider>
-                        <PriceProvider price={price}>
-                          <FireworksProvider>
-                            <ToastProvider>
-                              <ShowModalProvider>
-                                <BlockHeightProvider blockHeight={blockHeight}>
-                                  <ChainFeeProvider chainFee={chainFee}>
-                                    <ErrorBoundary>
-                                      <Component ssrData={ssrData} {...otherProps} />
-                                      {!router?.query?.disablePrompt && <PWAPrompt copyBody='This website has app functionality. Add it to your home screen to use it in fullscreen and receive notifications. In Safari:' promptOnVisit={2} />}
-                                    </ErrorBoundary>
-                                  </ChainFeeProvider>
-                                </BlockHeightProvider>
-                              </ShowModalProvider>
-                            </ToastProvider>
-                          </FireworksProvider>
-                        </PriceProvider>
-                      </ServiceWorkerProvider>
-                    </WebLnProvider>
-                  </LoggerProvider>
+                  <WebLnProvider>
+                    <ServiceWorkerProvider>
+                      <PriceProvider price={price}>
+                        <FireworksProvider>
+                          <ToastProvider>
+                            <ShowModalProvider>
+                              <BlockHeightProvider blockHeight={blockHeight}>
+                                <ChainFeeProvider chainFee={chainFee}>
+                                  <ErrorBoundary>
+                                    <Component ssrData={ssrData} {...otherProps} />
+                                    {!router?.query?.disablePrompt && <PWAPrompt copyBody='This website has app functionality. Add it to your home screen to use it in fullscreen and receive notifications. In Safari:' promptOnVisit={2} />}
+                                  </ErrorBoundary>
+                                </ChainFeeProvider>
+                              </BlockHeightProvider>
+                            </ShowModalProvider>
+                          </ToastProvider>
+                        </FireworksProvider>
+                      </PriceProvider>
+                    </ServiceWorkerProvider>
+                  </WebLnProvider>
                 </HasNewNotesProvider>
               </WalletsProvider>
             </MeProvider>
